requirements.txt
altair==5.0.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.5
Flask-WTF==1.1.1
gitdb==4.0.10
GitPython==3.1.32
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-auth==2.23.3
google-generativeai==0.2.1
googleapis-common-protos==1.61.0
greenlet==2.0.2
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.24.1
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
python-engineio==4.6.1
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
WTForms==3.0.1
zipp==3.16.2
